+++ terratex_reallife/SQLS/update.1.2.0.sql
@@ -1,3 +1,5 @@
 ALTER TABLE mta_terratex_reallife.user_gold ADD SkillBooster INT DEFAULT 0 NULL;
 
 ALTER TABLE mta_terratex_reallife.user ADD Gender VARCHAR(20) DEFAULT "anonym" NULL;
+
+ALTER TABLE mta_terratex_reallife.user_data ADD children VARCHAR(255) DEFAULT "[]" NULL;